+++ lib/provider/hive/application_settings.dart
@@ -45,11 +45,11 @@ class ApplicationSettingsHiveProvider
       putSafe(0, (box.get(0) ?? ApplicationSettings())..enablePopups = enabled);
 
   /// Stores a new [locale] value of [ApplicationSettings.locale] to [Hive].
-  Future<void> setLocale(String locale) async =>
-      await putSafe(0, (box.get(0) ?? ApplicationSettings())..locale = locale);
+  Future<void> setLocale(String locale) =>
+      putSafe(0, (box.get(0) ?? ApplicationSettings())..locale = locale);
 
-  /// Sets [ApplicationSettings.wasIntroductionShowed] flag.
-  Future<void> setIntroductionWasShowed(bool wasShowed) async => await putSafe(
-      0,
-      (box.get(0) ?? ApplicationSettings())..wasIntroductionShowed = wasShowed);
+  /// Stores a new [show] value of [ApplicationSettings.showIntroduction] to
+  /// [Hive].
+  Future<void> setIntroductionWasShowed(bool show) => putSafe(
+      0, (box.get(0) ?? ApplicationSettings())..showIntroduction = show);
 }