+++ lib/ui/page/auth/view.dart
@@ -21,11 +21,11 @@ import 'package:get/get.dart';
 
 import '/l10n/l10n.dart';
 import '/routes.dart';
-import '/ui/page/auth/login/view.dart';
+import '/ui/page/login/view.dart';
+import '/ui/widget/outlined_rounded_button.dart';
 import '/ui/widget/svg/svg.dart';
 import '/util/platform_utils.dart';
 import 'controller.dart';
-import 'widget/outlined_rounded_button.dart';
 
 /// View of the [Routes.auth] page.
 class AuthView extends StatelessWidget {