+++ lib/ui/page/home/tab/menu/controller.dart
@@ -24,9 +24,9 @@ import '/domain/service/call.dart';
 import '/domain/service/my_user.dart';
 import '/l10n/l10n.dart';
 import '/routes.dart';
-import '/ui/page/home/tab/menu/confirm/view.dart';
 import '/util/message_popup.dart';
 import '/util/web/web_utils.dart';
+import 'confirm/view.dart';
 
 export 'view.dart';
 
@@ -51,22 +51,23 @@ class MenuTabController extends GetxController {
   ///
   /// Shows a confirmation popup if there's any ongoing calls.
   Future<bool> confirmLogout() async {
-    bool value = true;
+    bool logout = true;
 
     if (_callService.calls.isNotEmpty || WebUtils.containsCalls()) {
       if (await MessagePopup.alert('alert_are_you_sure_want_to_log_out'.l10n) !=
           true) {
-        value = false;
+        logout = false;
       }
     }
 
+    // TODO: [MyUserService.myUser] might still be `null` there.
     if (_myUserService.myUser.value?.hasPassword != true) {
       if (await ConfirmLogoutView.show(router.context!) != true) {
-        value = false;
+        logout = false;
       }
     }
 
-    return value;
+    return logout;
   }
 
   /// Logs out the current session and go to the [Routes.auth] page.