+++ lib/ui/widget/context_menu/overlay.dart
@@ -48,15 +48,15 @@ class ContextMenuOverlay extends StatefulWidget {
 /// State of a [ContextMenuOverlay].
 class ContextMenuOverlayState extends State<ContextMenuOverlay> {
   /// Currently opened context menu.
-  final Rx<Widget?> _menu = Rx(null);
+  final Rx<Widget?> menu = Rx(null);
 
   /// Size of the [ContextMenuOverlay].
   Size? _area;
 
-  /// Position of the [_menu].
+  /// Position of the [menu].
   Offset _position = Offset.zero;
 
-  /// Alignment of this [_menu].
+  /// Alignment of this [menu].
   ///
   /// See [alignment] for details.
   Alignment _alignment = Alignment.bottomRight;
@@ -64,13 +64,11 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
   /// Alignment of current context menu.
   ///
   /// May be:
-  /// - `bottomRight`, meaning [_menu] is placed in the bottom right quadrant.
-  /// - `bottomLeft`, meaning [_menu] is placed in the bottom left quadrant.
-  /// - `topRight`, meaning [_menu] is placed in the top right quadrant.
-  /// - `topLeft`, meaning [_menu] is placed in the top left quadrant.
+  /// - `bottomRight`, meaning [menu] is placed in the bottom right quadrant.
+  /// - `bottomLeft`, meaning [menu] is placed in the bottom left quadrant.
+  /// - `topRight`, meaning [menu] is placed in the top right quadrant.
+  /// - `topLeft`, meaning [menu] is placed in the top left quadrant.
   Alignment get alignment => _alignment;
-
-  Rx<Widget?> get menu => _menu;
 
   @override
   Widget build(BuildContext context) {
@@ -80,7 +78,7 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
         // This is classic context menu behavior at the OS level.
         final Size size = constraints.biggest;
         if (size != _area) {
-          _menu.value = null;
+          menu.value = null;
           _area = size;
         }
 
@@ -98,12 +96,12 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
             body: Stack(
               children: [
                 widget.child,
-                if (_menu.value != null) ...[
-                  // Listens for taps outside the [_menu].
+                if (menu.value != null) ...[
+                  // Listens for taps outside the [menu].
                   Listener(
                     behavior: HitTestBehavior.opaque,
                     onPointerDown: (d) {
-                      // If [kSecondaryButton] was pressed outside the [_menu],
+                      // If [kSecondaryButton] was pressed outside the [menu],
                       // then simulate the [PointerUpDown] and [PointerUpEvent]
                       // for every [RenderPointerListener] on
                       // [BoxHitTestResult]'s path.
@@ -157,7 +155,7 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
                       hide();
                     },
                   ),
-                  // Draws [_menu] at [_position] with [alignment] translation.
+                  // Draws [menu] at [_position] with [alignment] translation.
                   Positioned(
                     left: _position.dx,
                     top: _position.dy,
@@ -166,7 +164,7 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
                         _alignment.x > 0 ? 0 : -1,
                         _alignment.y > 0 ? 0 : -1,
                       ),
-                      child: IntrinsicWidth(child: _menu.value),
+                      child: IntrinsicWidth(child: menu.value),
                     ),
                   ),
                 ]
@@ -181,11 +179,11 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
   /// Sets the current menu to [child] at [position].
   void show(Widget child, Offset position) => setState(() {
         _position = position;
-        _menu.value = child;
+        menu.value = child;
       });
 
   /// Hides the current menu if there is one.
-  void hide() => setState(() => _menu.value = null);
+  void hide() => setState(() => menu.value = null);
 }
 
 /// [InheritedWidget] of a [_ContextMenuOverlayState] used to implement