+++ CHANGELOG.md
@@ -16,15 +16,16 @@ All user visible changes to this project will be documented in this file. This p
 - UI:
     - User information auto-updating on changes ([#7], [#4]).
     - Menu:
-        - Language selection ([#23]).
+        - Language selection ([#23]);
+        - Unspecified password reminder on logout ([#35]).
+    - Introduction modal ([#35]).
 
 ### Changed
 
 - UI:
     - Media panel:
         - Redesigned desktop interface ([#26]).
-    - Login:
-        - Redesigned login interface ([#35])
+    - Redesigned login interface ([#35]).
 
 ### Fixed